online forecasting with real data improvement, last version was working OK

+++ mth_project/industrial_network_analysis/zabbix_simple/collect_data.py
@@ -115,7 +115,7 @@ class ZabbixDataCollector:
             
         try:
             time_to = int(time.time())
-            time_from = time_to - (hours_back * 3600)
+            time_from = int(time_to - (hours_back * 3600))  # Ensure integer timestamp
 
             print(f"Collecting {hours_back} hours of historical data...")
